create components for all elements, split as atomic design
