+++ src/components/blog/arrowRightIcon.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+function ArrowRightIcon(props) {
+    return (
+        <svg viewBox="0 0 16 16" fill="none" aria-hidden="true" {...props}>
+            <path
+                d="M7.25 4.75l3.5 3.25m0 0-3.5 3.25M11.75 8H3.25"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+            />
+        </svg>
+    )
+}
