+++ src/components/blog/socialLink.jsx
@@ -1,0 +1,10 @@
+import Link from "next/link";
+import React from "react";
+
+export function SocialLink({icon: Icon, ...props}) {
+    return (
+        <Link className="group -m-1 p-1" {...props}>
+            <Icon className="h-6 w-6 fill-zinc-500 transition group-hover:fill-zinc-600 dark:fill-zinc-400 dark:group-hover:fill-zinc-300"/>
+        </Link>
+    )
+}
