+++ src/pages/blog.jsx
@@ -1,237 +1,15 @@
 import Head from 'next/head'
-import Image from 'next/image'
-import Link from 'next/link'
-import clsx from 'clsx'
-import React, {useEffect} from 'react'
-
-import {Button} from '@/components/Button'
-import {Card} from '@/components/Card'
+import React from 'react'
 import {Container} from '@/components/Container'
 import {InstagramIcon, LinkedInIcon} from '@/components/SocialIcons'
-import logoPlanetaria from '@/images/planetaria.svg'
-import image1 from '@/images/photos/image-1.jpg'
-import image2 from '@/images/photos/image-2.jpg'
-import image3 from '@/images/photos/image-3.jpg'
-import image4 from '@/images/photos/image-4.jpg'
-import image5 from '@/images/photos/image-5.jpg'
-import {formatDate} from '@/lib/formatDate'
 import {Header} from '@/components/Header'
-import PocketBase from 'pocketbase'
-import loader from '@/components/Loader'
-import {log} from 'next/dist/server/typescript/utils'
-import Loader from "@/components/Loader";
+import Loader from '@/components/Loader'
 import {initPocketBase} from "../../lib/pocketbase";
 import {useQuery} from "@tanstack/react-query";
-
-function MailIcon(props) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            fill="none"
-            strokeWidth="1.5"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            aria-hidden="true"
-            {...props}
-        >
-            <path
-                d="M2.75 7.75a3 3 0 0 1 3-3h12.5a3 3 0 0 1 3 3v8.5a3 3 0 0 1-3 3H5.75a3 3 0 0 1-3-3v-8.5Z"
-                className="fill-zinc-100 stroke-zinc-400 dark:fill-zinc-100/10 dark:stroke-zinc-500"
-            />
-            <path
-                d="m4 6 6.024 5.479a2.915 2.915 0 0 0 3.952 0L20 6"
-                className="stroke-zinc-400 dark:stroke-zinc-500"
-            />
-        </svg>
-    )
-}
-
-function BriefcaseIcon(props) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            fill="none"
-            strokeWidth="1.5"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            aria-hidden="true"
-            {...props}
-        >
-            <path
-                d="M2.75 9.75a3 3 0 0 1 3-3h12.5a3 3 0 0 1 3 3v8.5a3 3 0 0 1-3 3H5.75a3 3 0 0 1-3-3v-8.5Z"
-                className="fill-zinc-100 stroke-zinc-400 dark:fill-zinc-100/10 dark:stroke-zinc-500"
-            />
-            <path
-                d="M3 14.25h6.249c.484 0 .952-.002 1.316.319l.777.682a.996.996 0 0 0 1.316 0l.777-.682c.364-.32.832-.319 1.316-.319H21M8.75 6.5V4.75a2 2 0 0 1 2-2h2.5a2 2 0 0 1 2 2V6.5"
-                className="stroke-zinc-400 dark:stroke-zinc-500"
-            />
-        </svg>
-    )
-}
-
-function ArrowRightIcon(props) {
-    return (
-        <svg viewBox="0 0 16 16" fill="none" aria-hidden="true" {...props}>
-            <path
-                d="M7.25 4.75l3.5 3.25m0 0-3.5 3.25M11.75 8H3.25"
-                strokeWidth="1.5"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-            />
-        </svg>
-    )
-}
-
-function Article({article}) {
-    return (
-        <Card as="article">
-            <Card.Title href={`/articles/${article.slug}`}>
-                {article.title}
-            </Card.Title>
-            <Card.Eyebrow as="time" dateTime={article.date} decorate>
-                {formatDate(article.date)}
-            </Card.Eyebrow>
-            <Card.Description>{article.description}</Card.Description>
-            <Card.Cta>Read article</Card.Cta>
-        </Card>
-    )
-}
-
-function SocialLink({icon: Icon, ...props}) {
-    return (
-        <Link className="group -m-1 p-1" {...props}>
-            <Icon className="h-6 w-6 fill-zinc-500 transition group-hover:fill-zinc-600 dark:fill-zinc-400 dark:group-hover:fill-zinc-300"/>
-        </Link>
-    )
-}
-
-function Newsletter() {
-    return (
-        <form
-            action="/thank-you"
-            className="rounded-2xl border border-zinc-100 p-6 dark:border-zinc-700/40"
-        >
-            <h2 className="flex text-sm font-semibold text-zinc-900 dark:text-zinc-100">
-                <MailIcon className="h-6 w-6 flex-none"/>
-                <span className="ml-3">Stay up to date</span>
-            </h2>
-            <p className="mt-2 text-sm text-zinc-600 dark:text-zinc-400">
-                Get notified when I publish something new, and unsubscribe at any time.
-            </p>
-            <div className="mt-6 flex">
-                <input
-                    type="email"
-                    placeholder="Email address"
-                    aria-label="Email address"
-                    required
-                    className="min-w-0 flex-auto appearance-none rounded-md border border-zinc-900/10 bg-white px-3 py-[calc(theme(spacing.2)-1px)] shadow-md shadow-zinc-800/5 placeholder:text-zinc-400 focus:border-teal-500 focus:outline-none focus:ring-4 focus:ring-teal-500/10 dark:border-zinc-700 dark:bg-zinc-700/[0.15] dark:text-zinc-200 dark:placeholder:text-zinc-500 dark:focus:border-teal-400 dark:focus:ring-teal-400/10 sm:text-sm"
-                />
-                <button
-                    type="submit"
-                    className="ml-4 inline-flex flex-none items-center justify-center gap-2
-                 rounded-md bg-zinc-800 px-3 py-2 text-sm
-                  font-semibold text-zinc-100 outline-offset-2 transition hover:bg-zinc-700 active:bg-zinc-800 active:text-zinc-100/70 active:transition-none dark:bg-zinc-700 dark:hover:bg-zinc-600 dark:active:bg-zinc-700 dark:active:text-zinc-100/70"
-                >
-                    Join
-                </button>
-            </div>
-        </form>
-    )
-}
-
-function Resume() {
-    let resume = [
-        {
-            company: 'v.01',
-            title: 'Première grosse mise à jour',
-            logo: logoPlanetaria,
-            start: '2019',
-            end: {
-                label: 'Present',
-                dateTime: new Date().getFullYear(),
-            },
-        },
-    ]
-
-    return (
-        <div className="rounded-2xl border border-zinc-100 p-6 dark:border-zinc-700/40">
-            <h2 className="flex text-sm font-semibold text-zinc-900 dark:text-zinc-100">
-                <BriefcaseIcon className="h-6 w-6 flex-none"/>
-                <span className="ml-3">Changelog</span>
-            </h2>
-            <ol className="mt-6 space-y-4">
-                {resume.map((role, roleIndex) => (
-                    <li key={roleIndex} className="flex gap-4">
-                        <div
-                            className="relative mt-1 flex h-10 w-10 flex-none items-center justify-center rounded-full shadow-md shadow-zinc-800/5 ring-1 ring-zinc-900/5 dark:border dark:border-zinc-700/50 dark:bg-zinc-800 dark:ring-0">
-                            <Image src={role.logo} alt="" className="h-7 w-7" unoptimized/>
-                        </div>
-                        <dl className="flex flex-auto flex-wrap gap-x-2">
-                            <dt className="sr-only">Company</dt>
-                            <dd className="w-full flex-none text-sm font-medium text-zinc-900 dark:text-zinc-100">
-                                {role.company}
-                            </dd>
-                            <dt className="sr-only">Role</dt>
-                            <dd className="text-xs text-zinc-500 dark:text-zinc-400">
-                                {role.title}
-                            </dd>
-                            <dt className="sr-only">Date</dt>
-                            <dd
-                                className="ml-auto text-xs text-zinc-400 dark:text-zinc-500"
-                                aria-label={`${role.start.label ?? role.start} until ${
-                                    role.end.label ?? role.end
-                                }`}
-                            >
-                                <time dateTime={role.start.dateTime ?? role.start}>
-                                    {role.start.label ?? role.start}
-                                </time>
-                                {' '}
-                                <span aria-hidden="true">—</span>{' '}
-                                <time dateTime={role.end.dateTime ?? role.end}>
-                                    {role.end.label ?? role.end}
-                                </time>
-                            </dd>
-                        </dl>
-                    </li>
-                ))}
-            </ol>
-            <Link
-                href="#"
-                className="group mt-6 inline-flex w-full items-center justify-center gap-2 rounded-md bg-zinc-50 px-3 py-2 text-sm font-medium text-zinc-900 outline-offset-2 transition hover:bg-zinc-100 active:bg-zinc-100 active:text-zinc-900/60 active:transition-none dark:bg-zinc-800/50 dark:text-zinc-300 dark:hover:bg-zinc-800 dark:hover:text-zinc-50 dark:active:bg-zinc-800/50 dark:active:text-zinc-50/70"
-            >
-                Voir les changelogs
-                <ArrowRightIcon className="h-4 w-4 stroke-zinc-400 transition group-active:stroke-zinc-600 dark:group-hover:stroke-zinc-50 dark:group-active:stroke-zinc-50"/>
-            </Link>
-        </div>
-    )
-}
-
-function Photos() {
-    let rotations = ['rotate-2', '-rotate-2', 'rotate-2', 'rotate-2', '-rotate-2']
-
-    return (
-        <div className="mt-16 sm:mt-20">
-            <div className="-my-4 flex justify-center gap-5 overflow-hidden py-4 sm:gap-8">
-                {[image1, image2, image3, image4, image5].map((image, imageIndex) => (
-                    <div
-                        key={image.src}
-                        className={clsx(
-                            'relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl',
-                            rotations[imageIndex % rotations.length]
-                        )}
-                    >
-                        <Image
-                            src={image}
-                            alt=""
-                            sizes="(min-width: 640px) 18rem, 11rem"
-                            className="absolute inset-0 h-full w-full object-cover"
-                        />
-                    </div>
-                ))}
-            </div>
-        </div>
-    )
-}
+import {Photos} from "@/components/blog/photos";
+import {Resume} from "@/components/blog/resume";
+import {Newsletter} from "@/components/blog/newsletter";
+import {SocialLink} from "@/components/blog/socialLink";
 
 /**
  * Blog page
